# feat: comprehensive performance optimization - 60-70% execution time reduction