… docs/COST_TIME_ANALYSIS.md
… docs/MADSPARK_OPTIMIZATION_PLAN.md
… src/madspark/utils/logical_inference_engine.py
… src/madspark/utils/output_processor.py
… test_logical_output.md
… tests/test_batch_logical_inference.py
… tests/test_export_logical_inference.py
… tests/test_logical_inference_formatting.py
… tests/test_parallel_processing.py